test(server): one phase-1 acceptance suite for the zoned .rdb
Folds the sidecar census (exit criterion (a)) and the migration round trip
(exit criterion (c)) into a single suite; they assert two halves of the same
clean break.

Refs #1960
